mslenses/requirements.txt
monai==1.5.1
antspyx==0.6.1
nibabel==5.3.2
numpy==1.26.4
nnunetv2==2.6.2
HD-BET==2.0.1